fix: resolve critical and high severity bugs
- Remove exposed API key and strengthen JWT secret
- Add null-safety checks in CurrentUserProvider
- Fix IDOR vulnerability in SemanticSearchController (add subject ownership check)
- Fix MaterialIngestionProcessor: mark failed on RAG error, don't rethrow RuntimeException
- Change character_count to BIGINT to prevent overflow
- Replace IVFFLAT with HNSW index for better small-dataset performance
- Remove redundant index on material_extracted_texts.material_id
- Fix integration test: persist Subject and Material entities before assertions
- Fix QuizIcon ReferenceError, replace with HelpCircle
- Remove broken /register and quizzes/new links
- Deduplicate cn() utility - import from @/lib/utils
- Fix minio-init race condition with service_healthy dependency

+++ backend/src/main/java/com/example/learning/materials/domain/MaterialExtractedText.java
@@ -24,7 +24,7 @@ public class MaterialExtractedText extends AuditableEntity {
     private Integer pageCount;
 
     @Column(name = "character_count", nullable = false)
-    private int characterCount;
+    private long characterCount;
 
     protected MaterialExtractedText() {
     }
@@ -53,7 +53,7 @@ public class MaterialExtractedText extends AuditableEntity {
         return pageCount;
     }
 
-    public int getCharacterCount() {
+    public long getCharacterCount() {
         return characterCount;
     }
